removed implemented method from DbReadNotificationRepository
