+++ centreon/src/Core/Notification/Infrastructure/API/FindNotificationsResources/FindNotificationsResourcesController.php
@@ -40,7 +40,7 @@ final class FindNotificationsResourcesController extends AbstractController
      * @param FindNotificationsResourcesPresenter $presenter
      * @param Request $request
      *
-     * @throws AccessDeniedException
+     * @throws AccessDeniedException|\Throwable
      *
      * @return Response
      */
