+++ centreon/src/Core/Notification/Infrastructure/Repository/HostGroupRequestProvider.php
@@ -21,11 +21,11 @@
 
 declare(strict_types=1);
 
-namespace Core\Notification\Infrastructure\Repository\NotifiableResourceRequestProvider;
+namespace Core\Notification\Infrastructure\Repository;
 
 use Core\Notification\Application\Repository\NotifiableResourceRequestProviderInterface;
 
-class ServiceGroupRequestProvider implements NotifiableResourceRequestProviderInterface
+class HostGroupRequestProvider implements NotifiableResourceRequestProviderInterface
 {
     /**
      * @inheritDoc
@@ -34,21 +34,21 @@ class ServiceGroupRequestProvider implements NotifiableResourceRequestProviderIn
     {
         return <<<'SQL'
             SELECT n.`id` AS `notification_id`,
-                hsr.`host_host_id` AS `host_id`,
+                h.`host_id` AS `host_id`,
                 h.`host_name` AS `host_name`,
                 h.`host_alias` AS `host_alias`,
                 n.`hostgroup_events` AS `host_events`,
-                s.`service_id` AS `service_id`,
+                hsr.`service_service_id` AS `service_id`,
                 s.`service_description` AS `service_name`,
                 s.`service_alias` AS `service_alias`,
-                n.`servicegroup_events` AS `service_events`,
-                0 AS `included_service_events`
-            FROM `:db`.`service` s
-                INNER JOIN `:db`.`servicegroup_relation` sgr ON sgr.`service_service_id` = s.`service_id`
-                INNER JOIN `:db`.`host_service_relation` hsr ON hsr.`service_service_id` = s.`service_id`
-                INNER JOIN `:db`.`notification_sg_relation` nsgr ON nsgr.`sg_id` = sgr.`servicegroup_sg_id`
-                INNER JOIN `:db`.`notification` n ON n.`id` = nsgr.`notification_id`
-                INNER JOIN `:db`.`host` h ON h.`host_id` = hsr.`host_host_id`
+                0 AS `service_events`,
+                n.`included_service_events`
+            FROM `:db`.`host` h
+                INNER JOIN `:db`.`hostgroup_relation` hgr ON hgr.`host_host_id` = h.`host_id`
+                INNER JOIN `:db`.`host_service_relation` hsr ON hsr.`host_host_id` = h.`host_id`
+                INNER JOIN `:db`.`notification_hg_relation` nhgr ON nhgr.`hg_id` = hgr.`hostgroup_hg_id`
+                INNER JOIN `:db`.`notification` n ON n.`id` = nhgr.`notification_id`
+                INNER JOIN `:db`.`service` s ON s.`service_id` = hsr.`service_service_id`
             WHERE n.`is_activated` = 1
             SQL;
     }
